New functions for packet package
Added Encapsulation and Decapsulation functions for packet:
EncapsulateHead, EncapsulateTail, DecapsulateHead, DecapsulateTail
which add or remove bytes inside a packet.

Added SwapBytesUint32, GetPacketLen, PacketBytesChange functions.
Made GetRawBytes function a member of packet package.
Made small type corrections.
Corresponding changes in low package and tests, examples.

+++ test/stability/test_merge/test-merge-part1.go
@@ -32,7 +32,7 @@ const (
 
 var (
 	// Payload is 16 byte md5 hash sum of headers
-	PAYLOAD_SIZE uint64 = 16
+	PAYLOAD_SIZE uint = 16
 
 	sentPacketsGroup1 uint64 = 0
 	sentPacketsGroup2 uint64 = 0
@@ -110,7 +110,7 @@ func main() {
 }
 
 func generatePacketGroup1(pkt *packet.Packet) {
-	packet.InitEmptyEtherIPv4UDPPacket(pkt, int(PAYLOAD_SIZE))
+	packet.InitEmptyEtherIPv4UDPPacket(pkt, PAYLOAD_SIZE)
 	if pkt == nil {
 		panic("Failed to create new packet")
 	}
@@ -126,7 +126,7 @@ func generatePacketGroup1(pkt *packet.Packet) {
 }
 
 func generatePacketGroup2(pkt *packet.Packet) {
-	packet.InitEmptyEtherIPv4UDPPacket(pkt, int(PAYLOAD_SIZE))
+	packet.InitEmptyEtherIPv4UDPPacket(pkt, PAYLOAD_SIZE)
 	if pkt == nil {
 		panic("Failed to create new packet")
 	}
