+++ test/stability/test_partition/test-partition-part1.go
@@ -38,7 +38,7 @@ const (
 
 var (
 	// Payload is 16 byte md5 hash sum of headers
-	PAYLOAD_SIZE uint64 = 16
+	PAYLOAD_SIZE uint = 16
 
 	sent          uint64     = 0
 	recvPackets   uint64     = 0
@@ -109,7 +109,7 @@ func main() {
 
 // Generate packets
 func generatePacket(pkt *packet.Packet) {
-	packet.InitEmptyEtherIPv4UDPPacket(pkt, int(PAYLOAD_SIZE))
+	packet.InitEmptyEtherIPv4UDPPacket(pkt, PAYLOAD_SIZE)
 	if pkt == nil {
 		panic("Failed to create new packet")
 	}
